Distribuimos os testes do ControllerTest.
