+++ src/test/java/model/acesso/PermissaoDAOTest.java
@@ -9,6 +9,8 @@ import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
 import db.DBConnection;
+import model.dao.PermissaoDAO;
+import model.entidades.Permissao;
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class PermissaoDAOTest {
